Resolve comments regarding syntax
- Not including changes for leaflet.d.ts type declaration for window.L
