+++ client/components/map/Marker/index.tsx
@@ -1,7 +1,6 @@
 // Marker Component to be dynamically imported
 import { Marker } from "react-leaflet";
 import useLeafletDivIcon from "hooks/useLeafletDivIcon";
-import useMapContext from "hooks/useMapContext";
 import MarkerIconWrapper, { CustomMarkerProps } from "./MarkerIconWrapper";
 
 export const CustomMarker: React.FC<{
@@ -11,13 +10,7 @@ export const CustomMarker: React.FC<{
 	handleOnClick: () => void;
 	category?: string;
 }> = ({ position, icon, color, handleOnClick, category }) => {
-	const { map } = useMapContext();
 	const { divIcon } = useLeafletDivIcon();
-
-	// const handleMarkerClick = () => {
-	// 	map?.panTo(position);
-	// 	handleOnClick();
-	// };
 
 	return (
 		<Marker
